Move to runtime.js
